+++ packages/ui/src/components/Range/index.ts
@@ -1,0 +1,1 @@
+export * from './Range';
